+++ src/types/models/article.ts
@@ -1,9 +1,9 @@
 import type { MicroCMSListContent, MicroCMSListResponse } from "microcms-js-sdk";
 
-import type { TCategory } from "@/types/category";
-import type { TTag } from "@/types/tag";
-import type { TImage, TImageOption } from "@/types/utils";
-import type { TWriter } from "@/types/writer";
+import type { TCategory } from "@/types/models/category";
+import type { TTag } from "@/types/models/tag";
+import type { TImage, TImageOption } from "@/types/models/utils";
+import type { TWriter } from "@/types/models/writer";
 
 export type TArticle = {
   image: TImage;
