+++ src/types/models/category.ts
@@ -1,6 +1,6 @@
 import type { MicroCMSListContent, MicroCMSListResponse } from "microcms-js-sdk";
 
-import type { TCategoryOrTagColor, TImage } from "@/types/utils";
+import type { TCategoryOrTagColor, TImage } from "@/types/models/utils";
 
 export type TCategory = {
   slug: string;
