+++ src/types/models/config.ts
@@ -1,6 +1,6 @@
 import type { MicroCMSObjectContent } from "microcms-js-sdk";
 
-import type { TImage } from "@/types/utils";
+import type { TImage } from "@/types/models/utils";
 
 export type TConfig = {
   host: string;
