+++ src/types/models/utils.ts
