+++ src/types/models/writer.ts
@@ -1,6 +1,6 @@
 import type { MicroCMSListContent, MicroCMSListResponse } from "microcms-js-sdk";
 
-import type { TImage } from "@/types/utils";
+import type { TImage } from "@/types/models/utils";
 
 export type TWriter = {
   name: string;
